requirements.txt
tensorflow-gpu==1.12
numpy
av
editdistance
configargparse
six

moviepy==0.2.3.5
opencv-python
